+++ src/main/java/com/eturn/eturn/service/impl/PositionServiceImpl.java
@@ -1,5 +1,6 @@
 package com.eturn.eturn.service.impl;
 
+import com.eturn.eturn.dto.MemberDTO;
 import com.eturn.eturn.dto.PositionDTO;
 import com.eturn.eturn.dto.PositionMoreInfoDTO;
 import com.eturn.eturn.dto.UserDTO;
@@ -7,6 +8,7 @@ import com.eturn.eturn.dto.mapper.PositionListMapper;
 import com.eturn.eturn.dto.mapper.PositionMapper;
 import com.eturn.eturn.dto.mapper.PositionMoreInfoMapper;
 import com.eturn.eturn.entity.*;
+import com.eturn.eturn.exception.position.NoAccessPosException;
 import com.eturn.eturn.exception.position.NoCreatePosException;
 import com.eturn.eturn.exception.position.NotFoundPosException;
 import com.eturn.eturn.repository.PositionRepository;
@@ -17,10 +19,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Calendar;
-import java.util.Date;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Stream;
 
 @Service
@@ -63,6 +62,12 @@ public class PositionServiceImpl implements PositionService {
                     List<Position> positionList = positionsPage.stream().toList();
                     Position p = positionList.get(positionList.size()-1);
                     positionRepository.tryToDelete(turn, p.getNumber());
+                    Optional<Position> positionFirstO2 = positionRepository.findFirstByTurnOrderByIdAsc(turn);
+                    if (positionFirstO2.isPresent()){
+                        Position positionF = positionFirstO2.get();
+                        positionF.setDateEnd(new Date());
+                        positionRepository.save(positionF);
+                    }
                 }
             }
         }
@@ -76,75 +81,80 @@ public class PositionServiceImpl implements PositionService {
         Turn turn = turnService.getTurnFrom(idTurn);
         UserDTO userDTO = userService.getUser(login);
         User user = userService.getUserFrom(userDTO.id());
+        MemberDTO memberDTO = memberService.getMember(user, turn);
+        if (!memberDTO.access().equals("BLOCKED")){
+            deleteOverdueElements(turn);
+            // рассчет участников
+            long members = memberService.getCountMembers(turn);
+            // если человек меньше, чем разрешенное число,
+            // то за него берется количество участников
+            int PERMITTED_COUNT_PEOPLE_SYSTEM = 2;
+            int PERMITTED_COUNT_PEOPLE = PERMITTED_COUNT_PEOPLE_SYSTEM;
+            if (members < PERMITTED_COUNT_PEOPLE_SYSTEM){
+                PERMITTED_COUNT_PEOPLE = (int) members;
+            }
+            // получение своей первой позиции
+            Optional<Position> ourPosition = positionRepository.findFirstByUserAndTurn(user, turn);
+            // Optional<Position> lastPositionInTurn = positionRepository.findTopByTurnOrderByNumberDesc(turn);
 
-        deleteOverdueElements(turn);
-        // рассчет участников
-        long members = memberService.getCountMembers(turn);
-        // если человек меньше, чем разрешенное число,
-        // то за него берется количество участников
-        int PERMITTED_COUNT_PEOPLE_SYSTEM = 2;
-        int PERMITTED_COUNT_PEOPLE = PERMITTED_COUNT_PEOPLE_SYSTEM;
-        if (members < PERMITTED_COUNT_PEOPLE_SYSTEM){
-            PERMITTED_COUNT_PEOPLE = (int) members;
-        }
-        // получение своей первой позиции
-        Optional<Position> ourPosition = positionRepository.findFirstByUserAndTurn(user, turn);
-        // Optional<Position> lastPositionInTurn = positionRepository.findTopByTurnOrderByNumberDesc(turn);
-
-        // получение списка из PERMITTED_COUNT_PEOPLE позиций для поиска нашей позиции
-        Pageable paging = PageRequest.of(0, PERMITTED_COUNT_PEOPLE);
-        Page<Position> positions = positionRepository.findByTurnOrderByIdDesc(turn, paging);
-        Stream<Position> allPositionsForEndStream = positions.get();
-        List<Position> allPositionsForEnd = allPositionsForEndStream.toList();
+            // получение списка из PERMITTED_COUNT_PEOPLE позиций для поиска нашей позиции
+            Pageable paging = PageRequest.of(0, PERMITTED_COUNT_PEOPLE);
+            Page<Position> positions = positionRepository.findByTurnOrderByIdDesc(turn, paging);
+            Stream<Position> allPositionsForEndStream = positions.get();
+            List<Position> allPositionsForEnd = allPositionsForEndStream.toList();
 //        long countAllPosForEnd = allPositionsForEnd.count();
 
-        // перебор списка
-        final int[] countBeforeUserPosFound = {0}; // количество позиций до появления позиции пользователя
-        final int[] userPosFound = {0}; // появление позиции пользователя
+            // перебор списка
+            final int[] countBeforeUserPosFound = {0}; // количество позиций до появления позиции пользователя
+            final int[] userPosFound = {0}; // появление позиции пользователя
 
-        if (!allPositionsForEnd.isEmpty()) {
-            allPositionsForEnd.forEach((element) -> {
-                if (element.getUser() == user && userPosFound[0] == 0) {
-                    userPosFound[0] = 1;
-                } else if (userPosFound[0] == 0) {
-                    countBeforeUserPosFound[0]++;
+            if (!allPositionsForEnd.isEmpty()) {
+                allPositionsForEnd.forEach((element) -> {
+                    if (element.getUser() == user && userPosFound[0] == 0) {
+                        userPosFound[0] = 1;
+                    } else if (userPosFound[0] == 0) {
+                        countBeforeUserPosFound[0]++;
+                    }
+                });
+            }
+            // проверка на то, что имеется позиция
+            boolean isUserPosExist = userPosFound[0]==1;
+
+            // разница для исключения
+            int differenceForException = PERMITTED_COUNT_PEOPLE - countBeforeUserPosFound[0];
+
+            if (isUserPosExist){
+                // если позиция существует, то делаем исключение
+                throw new NoCreatePosException(String.valueOf(differenceForException));
+            }
+            else{
+                // вычисляем номер позиции
+                int lastNumberPosition = 1;
+                if (!allPositionsForEnd.isEmpty()){
+                    Position lastPosition = allPositionsForEnd.get(0);
+                    lastNumberPosition = lastPosition.getNumber()+1;
                 }
-            });
-        }
-        // проверка на то, что имеется позиция
-        boolean isUserPosExist = userPosFound[0]==1;
 
-        // разница для исключения
-        int differenceForException = PERMITTED_COUNT_PEOPLE - countBeforeUserPosFound[0];
+                // создаем новую позицию
+                Position newPosition = new Position();
+                newPosition.setStart(false);
+                newPosition.setUser(user);
+                newPosition.setTurn(turn);
+                newPosition.setDateEnd(null);
+                newPosition.setGroupName(userDTO.group());
+                newPosition.setNumber(lastNumberPosition);
+                Position p = positionRepository.save(newPosition);
 
-        if (isUserPosExist){
-            // если позиция существует, то делаем исключение
-            throw new NoCreatePosException(String.valueOf(differenceForException));
+                // вычисляем разницу для позиции
+                int differenceForUser = -1;
+                if (ourPosition.isEmpty()){
+                    differenceForUser = (int) positionRepository.countNumbersLeft(p.getNumber(), turn);
+                }
+                return positionMoreInfoMapper.positionMoreInfoToPositionDTO(p, differenceForUser);
+            }
         }
         else{
-            // вычисляем номер позиции
-            int lastNumberPosition = 1;
-            if (!allPositionsForEnd.isEmpty()){
-                Position lastPosition = allPositionsForEnd.get(0);
-                lastNumberPosition = lastPosition.getNumber()+1;
-            }
-
-            // создаем новую позицию
-            Position newPosition = new Position();
-            newPosition.setStart(false);
-            newPosition.setUser(user);
-            newPosition.setTurn(turn);
-            newPosition.setDateEnd(null);
-            newPosition.setGroupName(userDTO.group());
-            newPosition.setNumber(lastNumberPosition);
-            Position p = positionRepository.save(newPosition);
-
-            // вычисляем разницу для позиции
-            int differenceForUser = -1;
-            if (ourPosition.isEmpty()){
-                differenceForUser = (int) positionRepository.countNumbersLeft(p.getNumber(), turn);
-            }
-            return positionMoreInfoMapper.positionMoreInfoToPositionDTO(p, differenceForUser);
+            throw new NoAccessPosException("You are blocked");
         }
     }
 
@@ -176,22 +186,31 @@ public class PositionServiceImpl implements PositionService {
     }
 
     @Override
-    public void update(Long id) {
+    public void update(Long id, String username) {
+        UserDTO userDTO = userService.getUser(username);
+        User user = userService.getUserFrom(userDTO.id());
         Optional<Position> position = positionRepository.findById(id);
         if (position.isPresent()){
             Position posI = position.get();
-            deleteOverdueElements(posI.getTurn());
-            Optional<Position> positionO = positionRepository.findById(id);
-            if (positionO.isPresent()){
-                Position pos = positionO.get();
-                if (pos.isStart()){
-                    delete(id);
-                }
-                else{
-                    pos.setStart(true);
-                    positionRepository.save(pos);
+            MemberDTO memberDTO = memberService.getMember(posI.getUser(), posI.getTurn());
+            if ((posI.getUser()==user && memberDTO.access().equals("MEMBER")) || memberDTO.access().equals("MODERATOR") || memberDTO.access().equals("CREATOR")){
+                deleteOverdueElements(posI.getTurn());
+                Optional<Position> positionO = positionRepository.findById(id);
+                if (positionO.isPresent()){
+                    Position pos = positionO.get();
+                    if (pos.isStart()){
+                        delete(id, pos.getUser().getUsername());
+                    }
+                    else{
+                        pos.setStart(true);
+                        positionRepository.save(pos);
+                    }
                 }
             }
+            else{
+                throw new NoAccessPosException("No access");
+            }
+
         }
         else{
             throw new NotFoundPosException("No positions found");
@@ -201,21 +220,30 @@ public class PositionServiceImpl implements PositionService {
 
     @Override
     @Transactional
-    public void delete(Long id) {
+    public void delete(Long id, String username) {
+        UserDTO userDTO = userService.getUser(username);
+        User user = userService.getUserFrom(userDTO.id());
         Optional<Position> position = positionRepository.findById(id);
         if (position.isPresent()) {
             Position pos = position.get();
-            positionRepository.delete(pos);
-            Optional<Position> p = positionRepository.findFirstByTurnOrderByIdAsc(pos.getTurn());
-            if (p.isPresent()){
-                Position changePosition = p.get();
-                Date date = new Date();
-                Calendar c = Calendar.getInstance();
-                c.setTime(date);
-                c.add(Calendar.MINUTE, 2);
-                changePosition.setDateEnd(c.getTime());
-                positionRepository.save(changePosition);
+            MemberDTO memberDTO = memberService.getMember(user, pos.getTurn());
+            if ((pos.getUser()==user && memberDTO.access().equals("MEMBER")) || memberDTO.access().equals("MODERATOR") || memberDTO.access().equals("CREATOR")){
+                positionRepository.delete(pos);
+                Optional<Position> p = positionRepository.findFirstByTurnOrderByIdAsc(pos.getTurn());
+                if (p.isPresent()){
+                    Position changePosition = p.get();
+                    Date date = new Date();
+                    Calendar c = Calendar.getInstance();
+                    c.setTime(date);
+                    c.add(Calendar.MINUTE, 2);
+                    changePosition.setDateEnd(c.getTime());
+                    positionRepository.save(changePosition);
+                }
             }
+            else{
+                throw new NoAccessPosException("No access");
+            }
+
         }
         else{
             throw new NotFoundPosException("No positions found");
@@ -225,7 +253,6 @@ public class PositionServiceImpl implements PositionService {
     @Override
     @Transactional
     public PositionMoreInfoDTO getFirstUserPosition(Long turnId, String username) {
-
 
         UserDTO userDTO = userService.getUser(username);
         User user = userService.getUserFrom(userDTO.id());
@@ -243,6 +270,30 @@ public class PositionServiceImpl implements PositionService {
             else{
                 int difference = (int) positionRepository.countNumbersLeft(p.get().getNumber(), turn);
                 return positionMoreInfoMapper.positionMoreInfoToPositionDTO(p.get(), difference);
+            }
+        }
+        else{
+            throw new NotFoundPosException("No positions found");
+        }
+    }
+
+    @Override
+    public PositionMoreInfoDTO getFirstPosition(Long turnId, String username) {
+        UserDTO userDTO = userService.getUser(username);
+        User user = userService.getUserFrom(userDTO.id());
+        Turn turn = turnService.getTurnFrom(turnId);
+
+        deleteOverdueElements(turn);
+        Optional<Position> pInTurn = positionRepository.findFirstByTurnOrderByIdAsc(turn);
+        if (pInTurn.isPresent()){
+            Position pos = pInTurn.get();
+            MemberDTO memberDTO = memberService.getMember(user, pos.getTurn());
+            if (memberDTO.access().equals("MODERATOR") || memberDTO.access().equals("CREATOR")){
+                int difference = 0;
+                return positionMoreInfoMapper.positionMoreInfoToPositionDTO(pos, difference);
+            }
+            else{
+                throw new NoAccessPosException("No access");
             }
         }
         else{
